step4 feat: 우승자 출력(step4 누락)
