+++ README.md
@@ -19,3 +19,5 @@
 - [x] 자동차는 0~9 사이의 랜덤값이 4 이상일 경우 전진한다
 - [x] 자동차는 0~9 사이의 랜덤값이 4 미만일 경우 멈춘다
 - [x] 자동차는 전진한 횟수만큼 position을 가진다
+- [] 포지션이 가장 큰 자동차가 우승한다
+- [] 우승자가 여러명일 경우 ,로 구분하여 출력한다
