+++ src/main/kotlin/racing/domain/CarNameMapper.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 object CarNameMapper {
     fun map(carNames: List<String>): List<Car> {
