+++ src/main/kotlin/racing/domain/Racing.kt
@@ -1,13 +1,14 @@
-package racing
+package racing.domain
 
-import racing.ui.ResultView
+import racing.view.ResultView
 
 class Racing(val cars: List<Car>, val tryCount: Int) {
     fun startRacing(movable: () -> Boolean): List<Car> {
         for (i in 0 until this.tryCount) {
             this.cars.filter { movable() }.forEach { it.move() }
+            ResultView.printResult(this.cars)
         }
-        ResultView.printResult(this.cars)
+        ResultView.printWinner(this.cars)
         return this.cars
     }
 }
