+++ src/test/kotlin/racing/CarNameMapperTest.kt
@@ -2,6 +2,7 @@ package racing
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
+import racing.domain.CarNameMapper
 
 class CarNameMapperTest : FunSpec({
     test("racing prepare시 position은 0으로 초기화()") {
