+++ src/test/kotlin/racing/CarTest.kt
@@ -5,6 +5,7 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.data.forAll
 import io.kotest.data.row
 import io.kotest.matchers.shouldBe
+import racing.domain.Car
 import java.lang.IllegalArgumentException
 
 class CarTest : FunSpec({
